# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVTable 后端服务 API",
        "title": "MVTable",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MVTable",
            "email": "mvtable@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/folder/all-nodes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "查询指定项目下的所有节点（文件夹、表格、仪表盘和文档），以树状结构返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "查询所有节点",
                "parameters": [
                    {
                        "type": "string",
                        "name": "projectId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QuerySubNodeRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/folder/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新的文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "创建文件夹",
                "parameters": [
                    {
                        "description": "创建文件夹请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateMvFolderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/folder/delete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "删除文件夹",
                "parameters": [
                    {
                        "description": "删除文件夹请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteMvFolderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/folder/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据项目ID查询该项目下所有文件夹，返回扁平结构列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "查询项目下所有文件夹（扁平列表）",
                "parameters": [
                    {
                        "type": "string",
                        "name": "projectId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功，返回扁平文件夹列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QueryFolderListRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/folder/move": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "将文件夹、数据表、仪表盘或文档移动到目标文件夹，targetFolderId为空字符串表示移动到根目录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "移动节点",
                "parameters": [
                    {
                        "description": "移动节点请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MoveNodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "移动成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/folder/search": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据关键字模糊搜索文件夹、表格、仪表盘与文档，返回树状结构",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "搜索文件夹、表格、仪表盘与文档",
                "parameters": [
                    {
                        "type": "string",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "projectId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功，返回树状结构",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SearchRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/folder/subquery": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "查询指定父节点下的子节点（文件夹、表格、仪表盘和文档）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "查询子节点",
                "parameters": [
                    {
                        "type": "string",
                        "name": "parentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QuerySubNodeRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/folder/update": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新文件夹信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件夹"
                ],
                "summary": "更新文件夹",
                "parameters": [
                    {
                        "description": "更新文件夹请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateMvFolderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middleware.Response": {
            "description": "统一响应结构",
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                },
                "state": {
                    "description": "响应状态码",
                    "type": "integer"
                }
            }
        },
...
        "model.CreateMvFolderReq": {
            "type": "object",
            "required": [
                "name",
                "projectId"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                }
            }
        },
...
        "model.DeleteMvFolderReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "model.FolderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                }
            }
        },
...
        "model.MoveNodeReq": {
            "type": "object",
            "required": [
                "projectId",
                "targetId",
                "type"
            ],
            "properties": {
                "projectId": {
                    "description": "项目ID",
                    "type": "string"
                },
                "targetFolderId": {
                    "description": "目标文件夹ID，为空字符串表示移动到根目录",
                    "type": "string"
                },
                "targetId": {
                    "description": "要移动的节点ID",
                    "type": "string"
                },
                "type": {
                    "description": "节点类型: folder / table / form / dashboard",
                    "type": "string"
                }
            }
        },
...
        "model.QueryFolderListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FolderItem"
                    }
                }
            }
        },
...
        "model.QuerySubNodeRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubNode"
                    }
                }
            }
        },
...
        "model.SearchRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubNode"
                    }
                }
            }
        },
...
        "model.SubNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubNode"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateMvFolderReq": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入Bearer token进行认证，格式: Bearer {token}。例如: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}